+++ src/main/java/carracing/domain/Car.java
@@ -3,27 +3,18 @@ package carracing.domain;
 public class Car {
     private String name;
     private int position;
-    private final String POSITION_MARKER = "-";
 
     public Car(String name){
+        this(name, 0);
+    }
+
+    public Car(String name, int position){
         this.name = name;
-        this.position = 0;
+        this.position = position;
     }
 
     public int go(){
         return this.position += 1;
-    }
-
-    public String getCarInfoString(){
-        StringBuilder positionBuilder = new StringBuilder();
-
-        positionBuilder.append(this.name + " : ");
-
-        for (int i = 0; i < this.position; i++) {
-            positionBuilder.append(POSITION_MARKER);
-        }
-
-        return positionBuilder.toString();
     }
 
     public String getName(){
