+++ src/main/java/carracing/view/CarRacingProgramPhrase.java
@@ -1,4 +1,4 @@
-package carracing.ui;
+package carracing.view;
 
 public class CarRacingProgramPhrase {
